refactor: move Hydration indexer and  config to dedicated module
